moved config header to component
