+++ src/components/crud-config.module.scss
@@ -4,39 +4,5 @@
     color: lightgrey;
 }
 
-.headerContainer {
-    display: flex;
-    padding: 8px 8px;
-    background-color: #2C3249;
-    align-items: center;
-    border-radius: 3px;
-}
 
-.detailesContainer {
-    display: flex;
-    flex-direction: column;
-
-    p,
-    span {
-        margin: 0;
-    }
-
-    span {
-        font-size: 12px;
-    }
-
-    p {
-        font-size: 14px;
-
-        span {
-            font-size: 14px;
-            color: gray;
-        }
-    }
-}
-
-
-.permissionsButtons {
-    margin-left: auto;
-}
